fix(storybook): move storybook back under src
