+++ package.json
@@ -19,7 +19,8 @@
   "license": "Apache-2.0",
   "files": [
     "**/*",
-    "!__tests__"
+    "!__tests__",
+    "!__stories__"
   ],
   "engines": {
     "node": ">=8.3.0"
