+++ src/__stories__/Text.tsx
@@ -5,7 +5,7 @@ import { boolean, select, text } from '@storybook/addon-knobs/react';
 import { storiesOf } from '@storybook/react';
 import { omitBy } from 'lodash';
 
-import { Text } from '../src/Text';
+import { Text } from '../Text';
 import { Casing, Decoration, LetterSpacing, LineHeight } from './_utils';
 import { boxKnobs } from './Box';
